% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef,SLGP-method}
\alias{coef,SLGP-method}
\title{Extract the coefficient draws of an SLGP model}
\usage{
\S4method{coef}{SLGP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SLGP-class}}.}

\item{...}{Ignored.}
}
\value{
The matrix of finite-rank GP coefficients (draws in rows, basis
functions in columns).
}
\description{
Extract the coefficient draws of an SLGP model
}
