% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{evaluate_basis_functions}
\alias{evaluate_basis_functions}
\title{Evaluate basis functions at given locations.}
\usage{
evaluate_basis_functions(parameters, X, lengthscale)
}
\arguments{
\item{parameters}{List of basis function parameters.}

\item{X}{Matrix or dataframe of evaluation locations.}

\item{lengthscale}{Numeric vector. Lengthscales used for scaling the input space.}
}
\value{
A matrix of basis function values.
}
\description{
Evaluates all basis functions defined by a parameter list at new locations.
}
\keyword{internal}
