% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{formula,SLGP-method}
\alias{formula,SLGP-method}
\title{Model formula of an SLGP object}
\usage{
\S4method{formula}{SLGP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{SLGP-class}}.}

\item{...}{Ignored.}
}
\value{
The model \code{formula}.
}
\description{
Model formula of an SLGP object
}
