% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{initialize_basisfun_fillingRFF}
\alias{initialize_basisfun_fillingRFF}
\title{Initialize space-filling Random Fourier Features}
\usage{
initialize_basisfun_fillingRFF(
  dimension,
  nFreq,
  MatParam = 5/2,
  lengthscale,
  seed = 0
)
}
\arguments{
\item{dimension}{Integer. Input (\eqn{[\mathbf{x},\,t]}{[x, t]}) dimension.}

\item{nFreq}{Integer. Number of frequency vectors to be considered.}

\item{MatParam}{Numeric. Matérn smoothness parameter (default = 5/2).}

\item{lengthscale}{Numeric vector. Lengthscales used for scaling the input space.}

\item{seed}{Integer. Random seed.}
}
\value{
List with frequency, offset, and coefficient parameters.
}
\description{
Initializes RFF parameters with LHS-optimized frequency directions.
}
\keyword{internal}
