% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{initialize_basisfun_RFF}
\alias{initialize_basisfun_RFF}
\title{Initialize parameters basis functions based on Random Fourier Features}
\usage{
initialize_basisfun_RFF(dimension, nFreq, MatParam = 5/2, lengthscale)
}
\arguments{
\item{dimension}{Integer. Input (\eqn{[\mathbf{x},\,t]}{[x, t]}) dimension.}

\item{nFreq}{Integer. Number of frequency vectors to be considered.}

\item{MatParam}{Numeric. Matérn smoothness parameter (default = 5/2).}

\item{lengthscale}{Numeric vector. Lengthscales used for scaling the input space.}
}
\value{
List with frequency, offset, and coefficient parameters.
}
\description{
Draws parameters for standard RFF approximating a Matérn kernel.
}
\keyword{internal}
