% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs,SLGP-method}
\alias{nobs,SLGP-method}
\title{Number of observations used to fit an SLGP model}
\usage{
\S4method{nobs}{SLGP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SLGP-class}}.}

\item{...}{Ignored.}
}
\value{
Integer number of observations.
}
\description{
Number of observations used to fit an SLGP model
}
