% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.SLGP}
\alias{print.summary.SLGP}
\title{Print method for SLGP summaries}
\usage{
\method{print}{summary.SLGP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.SLGP"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for SLGP summaries
}
